Clean Up
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@2069 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/plugins/optional/jdchat/Utils.java
@@ -7,7 +7,7 @@ import jd.utils.HTMLEntities;
 public class Utils {
     public static String getRandomColor() {
         
-        String col = Integer.toHexString((int)new Color((int) (Math.random() * 0xffffff)).darker().getRGB());
+        String col = Integer.toHexString(new Color((int) (Math.random() * 0xffffff)).darker().getRGB());
         while (col.length() < 6)
             col = "0" + col;
         return col.substring(col.length()-6);
